+++ src/pauli/dense.rs
@@ -10,6 +10,8 @@ use serde::{
     Serialize,
 };
 
+use super::Pauli;
+
 // just to effectively have an impl bool to make things more convenient here; the
 // disadvantage is that we cannot define the methods to be const but we don't need that
 trait ResolvePauli {
@@ -44,7 +46,7 @@ impl ResolvePauli for bool {
 /// possible to circumvent the invariant are unsafe.
 #[derive(Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Default, Debug)]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-pub struct Pauli {
+pub struct PauliDense {
     storage: u8,
 }
 
@@ -53,63 +55,31 @@ macro_rules! const_pauli {
         /// Encoded Pauli
         #[doc = $doc]
         /// .
-        pub const $name: Pauli = Pauli { storage: $value };
+        pub const $name: PauliDense = PauliDense { storage: $value };
     )*};
 }
-
 const_pauli!(
     (PAULI_I, 0, "I"),
     (PAULI_X, 2, "X"),
     (PAULI_Y, 3, "Y"),
-    (PAULI_Z, 1, "%"),
+    (PAULI_Z, 1, "Z"),
 );
 
-impl TryFrom<u8> for Pauli {
+impl TryFrom<u8> for PauliDense {
     type Error = u8;
     fn try_from(value: u8) -> Result<Self, Self::Error> {
         if value > 3 { Err(value) } else { Ok(Self { storage: value }) }
     }
 }
 
-impl From<Pauli> for u8 {
+impl From<PauliDense> for u8 {
     #[inline(always)]
-    fn from(value: Pauli) -> u8 {
+    fn from(value: PauliDense) -> u8 {
         value.storage
     }
 }
 
-macro_rules! new {
-    ($(($name:ident, $gate:ident),)*) => {$(
-        /// Create a new
-        #[doc = stringify!($gate)]
-        /// Pauli.
-        #[inline]
-        pub fn $name() -> Self {
-            // Safety: hardcoded in super::encoding
-            unsafe { Self::from_unchecked(super::encoding::$gate) }
-        }
-    )*};
-}
-
-impl Pauli {
-    /// Create a the new Pauli (X if x)(Z if z).
-    ///
-    /// # Examples
-    /// ```
-    /// # #[cfg_attr(coverage_nightly, no_coverage)]
-    /// # fn main() {
-    /// # use pauli_tracker::pauli::Pauli;
-    /// assert_eq!(Pauli::new(false, false), Pauli::new_i());
-    /// assert_eq!(Pauli::new(false, true), Pauli::new_z());
-    /// assert_eq!(Pauli::new(true, false), Pauli::new_x());
-    /// assert_eq!(Pauli::new(true, true), Pauli::new_y());
-    /// # }
-    pub fn new(x: bool, z: bool) -> Self {
-        Self { storage: x.left() ^ z.right() }
-    }
-
-    new!((new_i, I), (new_x, X), (new_y, Y), (new_z, Z),);
-
+impl PauliDense {
     /// Create a [Pauli] from a [u8] without checking the types invariant.
     ///
     /// # Safety
@@ -162,79 +132,7 @@ impl Pauli {
         self.storage = storage;
     }
 
-    /// Set whether the Pauli products contains X.
-    ///
-    /// # Examples
-    /// ```
-    /// # #[cfg_attr(coverage_nightly, no_coverage)]
-    /// # fn main() {
-    /// # use pauli_tracker::pauli::Pauli;
-    /// let mut pauli = Pauli::new_y();
-    /// pauli.set_x(false);
-    /// assert_eq!(pauli, Pauli::new_z());
-    /// # }
-    pub fn set_x(&mut self, x: bool) {
-        self.storage &= x.left() | 1;
-        self.storage |= x.left();
-    }
-
-    /// Set whether the Pauli products contains Z.
-    ///
-    /// # Examples
-    /// ```
-    /// # #[cfg_attr(coverage_nightly, no_coverage)]
-    /// # fn main() {
-    /// # use pauli_tracker::pauli::Pauli;
-    /// let mut pauli = Pauli::new_y();
-    /// pauli.set_z(false);
-    /// assert_eq!(pauli, Pauli::new_x());
-    /// # }
-    pub fn set_z(&mut self, z: bool) {
-        self.storage &= z.right() | 2;
-        self.storage |= z.right();
-    }
-
-    /// Get whether the Pauli products contains X.
-    ///
-    /// # Examples
-    /// ```
-    /// # #[cfg_attr(coverage_nightly, no_coverage)]
-    /// # fn main() {
-    /// # use pauli_tracker::pauli::Pauli;
-    /// let pauli = Pauli::new_y();
-    /// assert_eq!(pauli.get_x(), true);
-    /// # }
-    pub fn get_x(&self) -> bool {
-        self.storage & 2 != 0
-    }
-
-    /// Get whether the Pauli products contains Z.
-    ///
-    /// # Examples
-    /// ```
-    /// # #[cfg_attr(coverage_nightly, no_coverage)]
-    /// # fn main() {
-    /// # use pauli_tracker::pauli::Pauli;
-    /// let pauli = Pauli::new_y();
-    /// assert_eq!(pauli.get_z(), true);
-    /// # }
-    pub fn get_z(&self) -> bool {
-        self.storage & 1 != 0
-    }
-
-    /// Conjugate the Pauli with the Hadamard Gate ignoring phases.
-    pub fn h(&mut self) {
-        self.storage ^= (self.storage & 1) << 1;
-        self.storage ^= (self.storage & 2) >> 1;
-        self.storage ^= (self.storage & 1) << 1;
-    }
-    /// Conjugate the Pauli with the S Gate ignoring phases.
-    pub fn s(&mut self) {
-        self.storage ^= (self.storage & 2) >> 1;
-    }
-
     // is mask the correct word here?
-    // write examples
     /// Get the X mask of the encoded storage.
     ///
     /// # Examples
@@ -277,7 +175,7 @@ impl Pauli {
     }
 }
 
-impl Display for Pauli {
+impl Display for PauliDense {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self.storage {
             0 => write!(f, "I"),
@@ -289,6 +187,70 @@ impl Display for Pauli {
     }
 }
 
+impl Pauli for PauliDense {
+    fn new(x: bool, z: bool) -> Self {
+        Self { storage: x.left() ^ z.right() }
+    }
+
+    new_impl!();
+
+    #[inline]
+    fn add(&mut self, other: Self) {
+        self.xor(other);
+    }
+
+    fn h(&mut self) {
+        self.storage ^= (self.storage & 1) << 1;
+        self.storage ^= (self.storage & 2) >> 1;
+        self.storage ^= (self.storage & 1) << 1;
+    }
+
+    #[inline]
+    fn s(&mut self) {
+        self.storage ^= (self.storage & 2) >> 1;
+    }
+
+    #[inline]
+    fn xpx(&mut self, other: &Self) {
+        self.xor_u8(other.xmask());
+    }
+
+    #[inline]
+    fn xpz(&mut self, other: &Self) {
+        self.xor_u8(other.zmask() << 1);
+    }
+
+    #[inline]
+    fn zpx(&mut self, other: &Self) {
+        self.xor_u8(other.xmask() >> 1);
+    }
+
+    #[inline]
+    fn zpz(&mut self, other: &Self) {
+        self.xor_u8(other.zmask());
+    }
+
+    #[inline]
+    fn get_x(&self) -> bool {
+        self.storage & 2 != 0
+    }
+
+    #[inline]
+    fn get_z(&self) -> bool {
+        self.storage & 1 != 0
+    }
+
+    fn set_x(&mut self, x: bool) {
+        self.storage &= x.left() | 1;
+        self.storage |= x.left();
+    }
+
+    fn set_z(&mut self, z: bool) {
+        self.storage &= z.right() | 2;
+        self.storage |= z.right();
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use coverage_helper::test;
@@ -297,12 +259,12 @@ mod tests {
 
     #[test]
     fn set() {
-        type Action = fn(&mut Pauli, bool);
+        type Action = fn(&mut PauliDense, bool);
         const ACTIONS: [(Action, &str, [/* false, false */ [u8; 4]; 2]); 2] = [
-            (Pauli::set_x, "set_x", [[0, 1, 0, 1], [2, 3, 2, 3]]),
-            (Pauli::set_z, "set_z", [[0, 0, 2, 2], [1, 1, 3, 3]]),
+            (PauliDense::set_x, "set_x", [[0, 1, 0, 1], [2, 3, 2, 3]]),
+            (PauliDense::set_z, "set_z", [[0, 0, 2, 2], [1, 1, 3, 3]]),
         ];
-        let mut pauli = Pauli::new_i();
+        let mut pauli = PauliDense::new_i();
         for action in ACTIONS {
             for (flag, checks) in [false, true].into_iter().zip(action.2) {
                 for (input, check) in (0u8..).zip(checks) {