+++ src/tracker/live.rs
@@ -37,56 +37,78 @@ use crate::{
 // BitVec, ...
 #[derive(Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Default, Debug)]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-pub struct LiveVector {
-    inner: Vec<Pauli>,
+pub struct LiveVector<T> {
+    inner: Vec<T>,
 }
 
-impl From<Vec<Pauli>> for LiveVector {
-    fn from(value: Vec<Pauli>) -> Self {
+impl<T> From<Vec<T>> for LiveVector<T> {
+    fn from(value: Vec<T>) -> Self {
         Self { inner: value }
     }
 }
 
-impl From<LiveVector> for Vec<Pauli> {
-    fn from(value: LiveVector) -> Self {
+impl<T> From<LiveVector<T>> for Vec<T> {
+    fn from(value: LiveVector<T>) -> Self {
         value.inner
     }
 }
 
-impl AsRef<Vec<Pauli>> for LiveVector {
-    fn as_ref(&self) -> &Vec<Pauli> {
+impl<T> AsRef<Vec<T>> for LiveVector<T> {
+    fn as_ref(&self) -> &Vec<T> {
         &self.inner
     }
 }
 
-impl LiveVector {
+impl<T> LiveVector<T> {
     /// Returns a reference to an element at index. Returns [None] if out of bounds.
-    pub fn get(&self, bit: usize) -> Option<&Pauli> {
+    pub fn get(&self, bit: usize) -> Option<&T> {
         self.inner.get(bit)
     }
     /// Returns a mutable reference to an element at index. Returns [None] if out of
     /// bounds.
-    pub fn get_mut(&mut self, bit: usize) -> Option<&mut Pauli> {
+    pub fn get_mut(&mut self, bit: usize) -> Option<&mut T> {
         self.inner.get_mut(bit)
     }
 }
 
 macro_rules! single {
-    ($($name:ident),*) => {$(
+    ($($name:ident,)*) => {$(
         fn $name(&mut self, bit: usize) {
             unwrap_get_mut!(self.inner, bit, stringify!($name)).$name()
         }
     )*};
 }
 
+macro_rules! movements {
+    ($(($name:ident, $plus:ident, $set:ident),)*) => {$(
+        fn $name(&mut self, source: usize, destination: usize) {
+            let (s, d) =
+                unwrap_get_two_mut!(self.inner, source, destination, stringify!($name));
+            d.$plus(s);
+            s.$set(false);
+        }
+    )*};
+}
+
 /// Note that the inner storage type is basically a vector. Therefore, the it may
 /// contain buffer qubits, even though they were not explicitly initialized.
-impl Tracker for LiveVector {
-    type Stack = Pauli;
+impl<T> Tracker for LiveVector<T>
+where
+    T: Pauli + Clone,
+{
+    type Stack = T;
+    type Pauli = T;
+
+    movements!(
+        (move_x_to_x, xpx, set_x),
+        (move_x_to_z, zpx, set_x),
+        (move_z_to_x, xpz, set_z),
+        (move_z_to_z, zpz, set_z),
+    );
 
     fn init(num_bits: usize) -> Self {
         LiveVector {
-            inner: vec![Pauli::new_i(); num_bits],
+            inner: vec![T::new_i(); num_bits],
         }
     }
 
@@ -95,7 +117,7 @@ impl Tracker for LiveVector {
         match bit.cmp(&len) {
             Ordering::Less => Some(bit),
             Ordering::Equal => {
-                self.inner.push(Pauli::new_i());
+                self.inner.push(T::new_i());
                 None
             }
             Ordering::Greater => {
@@ -103,65 +125,40 @@ impl Tracker for LiveVector {
                 self.inner.try_reserve(diff).unwrap_or_else(|e| {
                     panic!("error when trying to reserve enough memory: {e}")
                 });
-                self.inner.extend(iter::repeat(Pauli::new_i()).take(diff));
+                self.inner.extend(iter::repeat(T::new_i()).take(diff));
                 None
             }
         }
     }
 
-    fn track_pauli(&mut self, bit: usize, pauli: Pauli) {
+    fn track_pauli(&mut self, bit: usize, pauli: T) {
         if let Some(p) = self.inner.get_mut(bit) {
-            p.xor(pauli)
+            p.add(pauli)
         }
     }
-    fn track_pauli_string(&mut self, string: PauliString) {
+    fn track_pauli_string(&mut self, string: PauliString<T>) {
         for (bit, pauli) in string {
             if let Some(p) = self.inner.get_mut(bit) {
-                p.xor(pauli)
+                p.add(pauli)
             }
         }
     }
 
-    single!(h, s);
+    single!(h, s,);
 
     fn cx(&mut self, control: usize, target: usize) {
         let (c, t) = unwrap_get_two_mut!(self.inner, control, target, "cx");
-        t.xor_u8(c.xmask());
-        c.xor_u8(t.zmask());
+        t.xpx(c);
+        c.zpz(t);
     }
     fn cz(&mut self, bit_a: usize, bit_b: usize) {
         let (a, b) = unwrap_get_two_mut!(self.inner, bit_a, bit_b, "cz");
-        a.xor_u8(b.xmask() >> 1);
-        b.xor_u8(a.xmask() >> 1);
-    }
-
-    fn move_x_to_x(&mut self, source: usize, destination: usize) {
-        let (s, d) =
-            unwrap_get_two_mut!(self.inner, source, destination, "move_x_to_x");
-        d.xor_u8(s.xmask());
-        s.set_x(false);
-    }
-    fn move_x_to_z(&mut self, source: usize, destination: usize) {
-        let (s, d) =
-            unwrap_get_two_mut!(self.inner, source, destination, "move_x_to_z");
-        d.xor_u8(s.xmask() >> 1);
-        s.set_x(false);
-    }
-    fn move_z_to_x(&mut self, source: usize, destination: usize) {
-        let (s, d) =
-            unwrap_get_two_mut!(self.inner, source, destination, "move_z_to_x");
-        d.xor_u8(s.zmask() << 1);
-        s.set_z(false);
-    }
-    fn move_z_to_z(&mut self, source: usize, destination: usize) {
-        let (s, d) =
-            unwrap_get_two_mut!(self.inner, source, destination, "move_z_to_z");
-        d.xor_u8(s.zmask());
-        s.set_z(false);
+        a.zpx(b);
+        b.zpx(a);
     }
 
     fn measure(&mut self, bit: usize) -> Result<Self::Stack, MissingStack> {
-        self.get(bit).ok_or(MissingStack { bit }).copied()
+        self.get(bit).ok_or(MissingStack { bit }).cloned()
     }
 }
 
@@ -170,52 +167,72 @@ mod tests {
     use coverage_helper::test;
 
     use super::*;
+    use crate::pauli::{
+        PauliDense,
+        PauliTuple,
+    };
 
-    mod action_definition_check {
-        use super::{
-            test,
-            *,
+    trait P: Pauli + Copy + Into<PauliDense> + From<PauliDense> {}
+
+    mod single_actions {
+        use super::*;
+        use crate::tracker::test::impl_utils::{
+            self,
+            SingleAction,
+            SingleResults,
+            N_SINGLES,
         };
+
+        type Action<T> = SingleAction<LiveVector<T>>;
+
+        #[cfg_attr(coverage_nightly, no_coverage)]
+        fn runner<T: P>(action: Action<T>, result: SingleResults) {
+            for (input, check) in (0u8..).zip(result.1) {
+                let mut tracker = LiveVector::init(2);
+                tracker.track_pauli_string(impl_utils::single_init(input));
+                (action)(&mut tracker, 0);
+                assert_eq!(
+                    T::into(*tracker.inner.get(0).unwrap()).storage(),
+                    check,
+                    "{}, {}",
+                    result.0,
+                    input
+                );
+            }
+        }
+
+        #[cfg_attr(coverage_nightly, no_coverage)]
+        pub(super) fn run<T: P>() {
+            let actions: [Action<T>; N_SINGLES] = [LiveVector::h, LiveVector::s];
+            impl_utils::single_check(runner, actions);
+        }
+    }
+
+    mod double_actions {
+        use super::*;
         use crate::tracker::test::impl_utils::{
             self,
             DoubleAction,
             DoubleResults,
-            SingleAction,
-            SingleResults,
             N_DOUBLES,
-            N_SINGLES,
         };
 
-        #[test]
-        fn single() {
-            type Action = SingleAction<LiveVector>;
+        type Action<T> = DoubleAction<LiveVector<T>>;
 
-            const ACTIONS: [Action; N_SINGLES] = [LiveVector::h, LiveVector::s];
-
-            #[cfg_attr(coverage_nightly, no_coverage)]
-            fn runner(action: Action, result: SingleResults) {
-                for (input, check) in (0u8..).zip(result.1) {
-                    let mut tracker = LiveVector::init(2);
-                    tracker.track_pauli_string(impl_utils::single_init(input));
-                    (action)(&mut tracker, 0);
-                    assert_eq!(
-                        tracker.inner.get(0).unwrap().storage(),
-                        check,
-                        "{}, {}",
-                        result.0,
-                        input
-                    );
-                }
+        #[cfg_attr(coverage_nightly, no_coverage)]
+        fn runner<T: P>(action: Action<T>, result: DoubleResults) {
+            for (input, check) in (0u8..).zip(result.1) {
+                let mut tracker = LiveVector::init(2);
+                tracker.track_pauli_string(impl_utils::double_init(input));
+                (action)(&mut tracker, 0, 1);
+                let output =
+                    impl_utils::double_output(tracker.inner.into_iter().enumerate());
+                assert_eq!(output, check, "{}, {}", result.0, input);
             }
-
-            impl_utils::single_check(runner, ACTIONS);
         }
 
-        #[test]
-        fn double() {
-            type Action = DoubleAction<LiveVector>;
-
-            const ACTIONS: [Action; N_DOUBLES] = [
+        pub(super) fn run<T: P>() {
+            let actions: [Action<T>; N_DOUBLES] = [
                 LiveVector::cx,
                 LiveVector::cz,
                 LiveVector::move_x_to_x,
@@ -224,22 +241,30 @@ mod tests {
                 LiveVector::move_z_to_z,
             ];
 
-            #[cfg_attr(coverage_nightly, no_coverage)]
-            fn runner(action: Action, result: DoubleResults) {
-                for (input, check) in (0u8..).zip(result.1) {
-                    let mut tracker = LiveVector::init(2);
-                    tracker.track_pauli_string(impl_utils::double_init(input));
-                    (action)(&mut tracker, 0, 1);
-                    let output = impl_utils::double_output(
-                        tracker.inner.into_iter().enumerate(),
-                    );
-                    assert_eq!(output, check, "{}, {}", result.0, input);
-                }
-            }
-
-            impl_utils::double_check(runner, ACTIONS);
+            impl_utils::double_check(runner, actions);
         }
     }
+
+    macro_rules! test_actions {
+        ($(($pauli:ty, $module:ident),)*) => {$(
+            mod $module {
+                use super::test;
+                #[rustfmt::skip]
+                use super::{double_actions, single_actions, P, $pauli};
+                impl P for $pauli {}
+                #[test]
+                fn single_actions() {
+                    single_actions::run::<$pauli>();
+                }
+                #[test]
+                fn double_actions() {
+                    double_actions::run::<$pauli>();
+                }
+            }
+        )*};
+    }
+
+    test_actions!((PauliDense, pauli_dense), (PauliTuple, pauli_tuple),);
 
     #[test]
     fn new_qubit_and_measure() {
@@ -247,12 +272,19 @@ mod tests {
         tracker.track_x(0);
         assert_eq!(tracker.new_qubit(0), Some(0));
         assert_eq!(tracker.new_qubit(1), None);
-        assert_eq!(*tracker.as_ref(), vec![Pauli::new_x(), Pauli::new_i()]);
-        assert_eq!(tracker.measure(0), Ok(Pauli::new_x()));
+        assert_eq!(*tracker.as_ref(), vec![PauliTuple::new_x(), PauliTuple::new_i()]);
+        assert_eq!(tracker.measure(0), Ok(PauliTuple::new_x()));
         assert_eq!(tracker.new_qubit(3), None);
         assert_eq!(
             *tracker.as_ref(),
-            vec![Pauli::new_x(), Pauli::new_i(), Pauli::new_i(), Pauli::new_i()]
+            vec![
+                PauliTuple::new_x(),
+                PauliTuple::new_i(),
+                PauliTuple::new_i(),
+                PauliTuple::new_i()
+            ]
         );
     }
+
+    //
 }